Fix Spring panel compilation errors in telemetry and governance audit maps

+++ spring-panel/src/main/java/com/example/panel/controller/AnalyticsController.java
@@ -403,7 +403,7 @@ public class AnalyticsController {
                 "analytics_legacy_inventory",
                 null,
                 "workspace.v1",
-                (double) scenarios.size(),
+                (long) scenarios.size(),
                 "workspace_v1_rollout",
                 null,
                 null,
